+++ integrations/storage/v1/definition/consts.go
@@ -1,0 +1,3 @@
+package definition
+
+const Name = "storage.v1"