+++ integrations/storage/v1/shared/s3/configuration.go
@@ -1,0 +1,11 @@
+package s3
+
+import (
+	awsHelper "github.com/arangodb/kube-arangodb/pkg/util/aws"
+)
+
+type Configuration struct {
+	BucketName string
+
+	Client awsHelper.Config
+}