+++ integrations/storage/v1/shared/s3/id.go
@@ -1,0 +1,12 @@
+package s3
+
+import (
+	"strings"
+
+	"github.com/dchest/uniuri"
+)
+
+// newID creates a new unique ID.
+func newID() string {
+	return strings.ToLower(uniuri.NewLen(uniuri.UUIDLen))
+}