+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - (Maintenance) PropagationMode Enum docs
 - (Feature) Deprecate AF Mode
 - (Maintenance) Switch License to 2025
+- (Feature) Migrate Storage V1 to CE
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD